fix type issue

+++ robust_speech/adversarial/filters/ASNRWiener.py
@@ -24,7 +24,7 @@ class ASNRWiener:
 
     def __call__(self, x: np.ndarray):
         print("ASNR filter launched ", x.shape)
-        x_filtered=np.copy(x.cpu().detach().numpy())
+        x_filtered=np.copy(x.cpu().detach().numpy().astype(float))
         for i in range(len(x)):
             if self.high_freq:
                 noise = np.random.normal(0, scale=self.gaussian_sigma, size=(x[i].shape[0]+1,))
